refactor: Move StatusConfiguration to its own file
