+++ tests/Commands/ToggleDone.test.ts
@@ -6,7 +6,8 @@ import moment from 'moment';
 import { calculateCursorOffset, toggleLine } from '../../src/Commands/ToggleDone';
 import { resetSettings, updateSettings } from '../../src/Config/Settings';
 import { StatusRegistry } from '../../src/StatusRegistry';
-import { Status, StatusConfiguration } from '../../src/Status';
+import { Status } from '../../src/Status';
+import { StatusConfiguration } from '../../src/StatusConfiguration';
 
 window.moment = moment;
